refactor: prune 10 superseded tools (34 -> 25); shared parsing to data_common.go
