+++ CLAUDE.md
@@ -389,10 +389,11 @@ fpl-draft-mcp/
 │   │   │   ├── dev/                     # the ONLY component that hits the live FPL API
 │   │   │   └── schema-inventory/        # dev utility: dumps API schema registry
 │   │   └── fpl-server/
-│   │       ├── main.go                  # Entry point, registers all 34 tools, auth, /mcp
+│   │       ├── main.go                  # Entry point, registers all 25 tools, auth, /mcp
 │   │       ├── draft_tools.go           # Decision layer: draft_board, player_card, waiver_plan, my_week, drop_radar (serve ML artifacts)
 │   │       ├── season_tools.go          # Decision layer: trade_check, league_pulse, team_env
-│   │       ├── waiver_recommendations.go# Waiver scoring logic
+│   │       ├── gw_live.go               # Decision layer: live H2H matchup tracker
+│   │       ├── data_common.go           # Shared bootstrap/live-GW parsing + points-conceded aggregation
 │   │       ├── fixture_difficulty.go    # FDR calculations
 │   │       ├── head_to_head.go          # H2H record tool
 │   │       ├── manager_season.go        # Season stats tool
@@ -401,9 +402,7 @@ fpl-draft-mcp/
 │   │       ├── player_gw_stats.go       # Per-GW player stats tool
 │   │       ├── current_roster.go        # Active roster tool
 │   │       ├── draft_picks.go           # Draft history tool
-│   │       ├── transaction_analysis.go  # Transaction ranking tool
-│   │       ├── league_entries.go        # League entries tool
-│   │       ├── epl_*.go / game_status.go# Global EPL data tools (standings, fixtures, status)
+│   │       ├── epl_*.go                 # Global EPL data tools (standings, fixtures)
 │   │       └── *_test.go                # Go unit tests (no live calls)
 │   └── backend/             # Python FastAPI backend (port 8000)
 │       └── backend/
@@ -443,7 +442,7 @@ Derive (Python, backend/ml/*)
   ▼
 Go MCP Server (:8080)
   │  reads raw + derived (local JSON only)
-  │  exposes 34 tools via MCP protocol (X-API-Key auth)
+  │  exposes 25 tools via MCP protocol (X-API-Key auth)
   ▼
 Claude Desktop / Claude Code (the LLM client, user's Max plan)
   │  calls decision tools, layers live web research + judgment
