+++ README.md
@@ -1,7 +1,7 @@
 # FPL Draft Co-Pilot
 
 A **draft + weekly-manager co-pilot** for Fantasy Premier League Draft.
-A Go MCP server exposes 34 tools over locally-cached FPL Draft API data; a
+A Go MCP server exposes 25 tools over locally-cached FPL Draft API data; a
 Python ML pipeline turns seven seasons of history into projections and
 weekly recommendations; **Claude (Desktop or Code) is the client** — you ask
 questions in natural language, Claude calls the decision tools and layers
@@ -9,7 +9,7 @@ live web research on top.
 
 ```
 FPL API → Go fetcher → data/<season>/ → Python ML (projections, waivers,
-start/sit) → Go MCP server (:8080, 34 tools) → Claude Desktop / Claude Code
+start/sit) → Go MCP server (:8080, 25 tools) → Claude Desktop / Claude Code
 ```
 
 ## The decision layer
@@ -24,8 +24,9 @@ start/sit) → Go MCP server (:8080, 34 tools) → Claude Desktop / Claude Code
 | `league_pulse` | What's happening? Standings, named transactions, game clock |
 | `drop_radar` | Who hit the wire? Ownership diffs from element-status snapshots |
 | `team_env` | Shootout or stalemate? Per-team points/xG generated and conceded, by position and venue |
+| `gw_live` | How's my matchup going? Live H2H tracker: both XIs with in-play points (refresh mid-match) |
 
-Plus the 26 data-layer tools (standings, matchups, fixtures, transactions,
+Plus 16 data-layer tools (matchups, fixtures, form, rosters,
 per-GW player stats, …) — all season-aware: flat `data/` roots are the
 2025-26 archive, current seasons nest under `data/{raw,derived}/<season>/`.
 
@@ -102,7 +103,7 @@ claude mcp add fpl --transport http http://localhost:8080/mcp \
 ```
 apps/
   mcp-server/            Go module
-    fpl-server/          34 MCP tool handlers + HTTP server (X-API-Key auth)
+    fpl-server/          25 MCP tool handlers + HTTP server (X-API-Key auth)
     cmd/dev/             FPL data fetcher (the only live-API component)
     internal/            fetch, store, ledger, points, summary, config
   backend/               Python package
