+++ apps/mcp-server/fpl-server/new_tools_test.go
@@ -745,181 +745,4 @@ func TestBuildPlayerGWStats(t *testing.T) {
 
 // ---- TestBuildTxRanking ----
 
-func TestBuildTxRanking(t *testing.T) {
-	playerByID := map[int]elementInfo{
-		1: {ID: 1, Name: "Salah", TeamID: 10, PositionType: 3},
-		2: {ID: 2, Name: "Haaland", TeamID: 11, PositionType: 4},
-		3: {ID: 3, Name: "TAA", TeamID: 10, PositionType: 2},
-	}
-	teamShort := map[int]string{10: "LIV", 11: "MCI"}
-
-	t.Run("SortedByCountDesc", func(t *testing.T) {
-		counts := map[int]int{1: 3, 2: 5, 3: 1}
-		out := buildTxRanking(counts, playerByID, teamShort, 10)
-		if out[0].Element != 2 {
-			t.Errorf("first=%d want 2 (count=5)", out[0].Element)
-		}
-		if out[1].Element != 1 {
-			t.Errorf("second=%d want 1 (count=3)", out[1].Element)
-		}
-		if out[2].Element != 3 {
-			t.Errorf("third=%d want 3 (count=1)", out[2].Element)
-		}
-	})
-
-	t.Run("LimitRespected", func(t *testing.T) {
-		counts := map[int]int{1: 3, 2: 5, 3: 1}
-		out := buildTxRanking(counts, playerByID, teamShort, 2)
-		if len(out) != 2 {
-			t.Errorf("len=%d want 2", len(out))
-		}
-	})
-
-	t.Run("TieBreakByLowerID", func(t *testing.T) {
-		counts := map[int]int{1: 3, 2: 3}
-		out := buildTxRanking(counts, playerByID, teamShort, 10)
-		// On equal count, lower element ID should come first.
-		if out[0].Element != 1 {
-			t.Errorf("first=%d want 1 (lower id on tie)", out[0].Element)
-		}
-	})
-
-	t.Run("EmptyCounts", func(t *testing.T) {
-		out := buildTxRanking(map[int]int{}, playerByID, teamShort, 10)
-		if len(out) != 0 {
-			t.Errorf("len=%d want 0 for empty counts", len(out))
-		}
-	})
-}
-
 // ---- TestBuildTransactionAnalysis ----
-
-func TestBuildTransactionAnalysis(t *testing.T) {
-	twoEntries := []any{
-		map[string]any{"id": 1, "entry_id": 200, "entry_name": "Alpha FC", "short_name": "AFC"},
-		map[string]any{"id": 2, "entry_id": 201, "entry_name": "Beta FC", "short_name": "BFC"},
-	}
-
-	t.Run("FiltersUnapprovedTransactions", func(t *testing.T) {
-		dir, cfg := tmpCfg(t)
-		writeBootstrap(t, dir)
-		writeLeagueDetailsFixture(t, dir, 100, twoEntries, nil)
-		writeJSON(t, filepath.Join(dir, "league/100/transactions.json"), map[string]any{
-			"transactions": []any{
-				// Approved waiver → included.
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 2, "event": 26, "kind": "w", "result": "a"},
-				// Not approved (result="d") → excluded.
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 2, "event": 26, "kind": "w", "result": "d"},
-				// Wrong GW → excluded.
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 2, "event": 25, "kind": "w", "result": "a"},
-				// Unknown kind → excluded.
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 2, "event": 26, "kind": "x", "result": "a"},
-			},
-		})
-		out, err := buildTransactionAnalysis(cfg, TransactionAnalysisArgs{LeagueID: 100, GW: 26})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if out.TotalTransactions != 1 {
-			t.Errorf("total=%d want 1 (only approved waiver in correct GW)", out.TotalTransactions)
-		}
-	})
-
-	t.Run("FreeAgentTransactionIncluded", func(t *testing.T) {
-		dir, cfg := tmpCfg(t)
-		writeBootstrap(t, dir)
-		writeLeagueDetailsFixture(t, dir, 100, twoEntries, nil)
-		writeJSON(t, filepath.Join(dir, "league/100/transactions.json"), map[string]any{
-			"transactions": []any{
-				// kind="f" (free agent) should be included.
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 2, "event": 26, "kind": "f", "result": "a"},
-			},
-		})
-		out, err := buildTransactionAnalysis(cfg, TransactionAnalysisArgs{LeagueID: 100, GW: 26})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if out.TotalTransactions != 1 {
-			t.Errorf("total=%d want 1 (free agent should be included)", out.TotalTransactions)
-		}
-	})
-
-	t.Run("PositionBreakdown", func(t *testing.T) {
-		// Add Salah (MID=3), drop Haaland (FWD=4).
-		dir, cfg := tmpCfg(t)
-		writeBootstrap(t, dir)
-		writeLeagueDetailsFixture(t, dir, 100, twoEntries, nil)
-		writeJSON(t, filepath.Join(dir, "league/100/transactions.json"), map[string]any{
-			"transactions": []any{
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 2, "event": 26, "kind": "f", "result": "a"},
-			},
-		})
-		out, err := buildTransactionAnalysis(cfg, TransactionAnalysisArgs{LeagueID: 100, GW: 26})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if out.PositionBreakdown["MID"].Added != 1 {
-			t.Errorf("MID.Added=%d want 1", out.PositionBreakdown["MID"].Added)
-		}
-		if out.PositionBreakdown["FWD"].Dropped != 1 {
-			t.Errorf("FWD.Dropped=%d want 1", out.PositionBreakdown["FWD"].Dropped)
-		}
-		if out.PositionBreakdown["DEF"].Added != 0 {
-			t.Errorf("DEF.Added=%d want 0", out.PositionBreakdown["DEF"].Added)
-		}
-	})
-
-	t.Run("TopAddedRanking", func(t *testing.T) {
-		// Salah added by both teams; Haaland added once.
-		dir, cfg := tmpCfg(t)
-		writeBootstrap(t, dir)
-		writeLeagueDetailsFixture(t, dir, 100, twoEntries, nil)
-		writeJSON(t, filepath.Join(dir, "league/100/transactions.json"), map[string]any{
-			"transactions": []any{
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 3, "event": 26, "kind": "w", "result": "a"},
-				map[string]any{"entry": 201, "element_in": 1, "element_out": 3, "event": 26, "kind": "f", "result": "a"},
-				map[string]any{"entry": 200, "element_in": 2, "element_out": 3, "event": 26, "kind": "w", "result": "a"},
-			},
-		})
-		out, err := buildTransactionAnalysis(cfg, TransactionAnalysisArgs{LeagueID: 100, GW: 26})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(out.TopAdded) == 0 {
-			t.Fatal("top_added is empty")
-		}
-		if out.TopAdded[0].Element != 1 {
-			t.Errorf("top added element=%d want 1 (Salah, count=2)", out.TopAdded[0].Element)
-		}
-		if out.TopAdded[0].Count != 2 {
-			t.Errorf("top added count=%d want 2", out.TopAdded[0].Count)
-		}
-	})
-
-	t.Run("ManagerActivityGroupedByEntry", func(t *testing.T) {
-		dir, cfg := tmpCfg(t)
-		writeBootstrap(t, dir)
-		writeLeagueDetailsFixture(t, dir, 100, twoEntries, nil)
-		writeJSON(t, filepath.Join(dir, "league/100/transactions.json"), map[string]any{
-			"transactions": []any{
-				map[string]any{"entry": 200, "element_in": 1, "element_out": 2, "event": 26, "kind": "w", "result": "a"},
-				map[string]any{"entry": 201, "element_in": 2, "element_out": 3, "event": 26, "kind": "f", "result": "a"},
-			},
-		})
-		out, err := buildTransactionAnalysis(cfg, TransactionAnalysisArgs{LeagueID: 100, GW: 26})
-		if err != nil {
-			t.Fatal(err)
-		}
-		if len(out.ManagerActivity) != 2 {
-			t.Errorf("manager_activity len=%d want 2", len(out.ManagerActivity))
-		}
-	})
-
-	t.Run("MissingLeagueID", func(t *testing.T) {
-		_, cfg := tmpCfg(t)
-		_, err := buildTransactionAnalysis(cfg, TransactionAnalysisArgs{})
-		if err == nil {
-			t.Fatal("expected league_id error")
-		}
-	})
-}
