+++ docs/CLAUDE_DESKTOP.md
@@ -1,9 +1,12 @@
 # Connecting Claude Desktop (or Claude Code) to the FPL co-pilot
 
-The Go MCP server serves 34 tools over Streamable HTTP at `/mcp`, including the
+The Go MCP server serves 25 tools over Streamable HTTP at `/mcp`, including the
 decision layer: `draft_board`, `player_card`, `waiver_plan`, `my_week`,
-`trade_check`, `league_pulse`, `drop_radar`, `team_env`. Claude (on a Max
-plan) is the client — there is no in-app LLM.
+`trade_check`, `league_pulse`, `drop_radar`, `team_env`, `gw_live`. Claude
+(on a Max plan) is the client — there is no in-app LLM.
+
+During matches: run a full (non `--fast`) fetch refresh, then ask
+*"gw_live — how's my matchup going?"* for both XIs with in-play points.
 
 ## 0. One-time setup: `.env` at the repo root
